Refactored platform
