+++ index.js
@@ -1,5 +1,5 @@
 
-const Platform = require('./src/platform_temp');
+const Platform = require('./src/platform');
 
 module.exports = function (homebridge) {
     homebridge.registerPlatform("homebridge-appletv-now-playing", "AppleTvNowPlayingPlatform", Platform, true);
