+++ src/platform.js
@@ -1,60 +1,82 @@
 const appletv = require("node-appletv-x");
 const Device = require("./device");
 
-function Platform(log, config, api) {
-    this.log = log;
-    this.config = config;
-    this.api = api;
-
-    this.accessories = [];
-    this.devices = [];
-
-    this.debug = debug.bind(this);
-    this.registerAccessories = registerAccessories.bind(this);
-    this.unregisterAccessories = unregisterAccessories.bind(this);
-    this.updateAccessories = updateAccessories.bind(this);
-    this.configureAccessory = configureAccessory.bind(this);
-    this.apiDidFinishLaunching = apiDidFinishLaunching.bind(this);
-    this.onScanComplete = onScanComplete.bind(this);
-    this.scanForDevice = scanForDevice.bind(this);
-    this.onDeviceConnected = onDeviceConnected.bind(this);
-    this.onDeviceConnectionFailed = onDeviceConnectionFailed.bind(this);
-    this.cleanupAccessory = cleanupAccessory.bind(this);
-
-    function registerAccessories(accessories) {
-        this.api.registerPlatformAccessories(Platform.pluginName, Platform.platformName, accessories);
+class Platform {
+    constructor(log, config, api) {
+        this.log = log;
+        this.config = config;
+        this.api = api;
+        this.accessories = [];
+        this.devices = [];
+        
+        this.api.on("didFinishLaunching", this.onApiDidFinishLaunching);
     }
 
-    function unregisterAccessories(accessories) {
-        this.api.unregisterPlatformAccessories(Platform.pluginName, Platform.platformName, accessories);
-    }
-
-    function updateAccessories(accessories) {
-        this.api.updatePlatformAccessories(accessories);
-    }
-
-    function debug(message) {
+    debug = message => {
         if (this.config && this.config.debug) {
             this.log(message);
         }
     }
 
-    function configureAccessory(accessory) {
+    registerAccessories = accessories => {
+        this.api.registerPlatformAccessories(Platform.pluginName, Platform.platformName, accessories);
+    }
+
+    unregisterAccessories = accessories => {
+        this.api.unregisterPlatformAccessories(Platform.pluginName, Platform.platformName, accessories);
+    }
+
+    updateAccessories = accessories => {
+        this.api.updatePlatformAccessories(accessories);
+    }
+
+    configureAccessory = accessory => {
         if (!accessory.context.uid) {
             this.debug(`Removing cached accessory width id ${accessory.UUID}`);
 
             this.api.unregisterPlatformAccessories(Platform.pluginName, Platform.platformName, [accessory]);
-        } else {
+        }
+        else {
             this.accessories.push(accessory);
 
             this.debug(`Loaded cached accessory width id ${accessory.UUID}`);
         }
     }
 
-    function apiDidFinishLaunching() {
-        if (!this.config.devices) {
-            this.debug("No devices have bene configured.");
+    cleanupAccessory = accessory => {
+        let foundAccessory = this.config.devices.filter(deviceConfiguration => {
+            let credentials = appletv.parseCredentials(deviceConfiguration.credentials);
+            accessory.UUID === `${credentials.uniqueIdentifier}_apple_tv`;
+        });
 
+        if (!foundAccessory) {
+            this.debug(`Removing orphaned accessory [${accessory.uid}].`);
+
+            this.unregisterAccessories([accessory]);
+        }
+    }
+
+    loadDevice = async deviceConfiguration => {
+        let credentials = appletv.parseCredentials(deviceConfiguration.credentials);
+
+        this.debug(`Scanning for Apple TV [${credentials.uniqueIdentifier}].`);
+
+        let devices = await appletv.scan(credentials.uniqueIdentifier);
+
+        this.debug(`Apple TV [${credentials.uniqueIdentifier}] found.`);
+        this.debug(`Attempting to connect to Apple TV [${credentials.uniqueIdentifier}].`);
+
+        let connectedDevice = await devices[0].openConnection(credentials);
+
+        this.debug(`Connected to ${connectedDevice.name} [${connectedDevice.uid}].`);
+        this.debug(`Loading acessory for ${connectedDevice.name} [${connectedDevice.uid}].`);
+
+        this.devices.push(new Device(this, deviceConfiguration, connectedDevice));
+    }
+
+    onApiDidFinishLaunching = () => {
+        if (!this.config.devices) {
+            this.debug("No Apple TV devices have been configured.");
             return;
         }
 
@@ -62,58 +84,10 @@ function Platform(log, config, api) {
 
         this.accessories.map(this.cleanupAccessory);
 
-        this.debug("Scanning for Apple TVs...");
-
-        this.config.devices.map(this.scanForDevice);
+        this.debug("Loading configured Apple TVs...");
+        
+        this.config.devices.map(this.loadDevice);
     }
-
-    function cleanupAccessory(accessory) {
-        let foundAccessory = this.config.devices.filter(deviceConfiguration => {
-          let credentials = appletv.parseCredentials(deviceConfiguration.credentials);
-          accessory.UUID === `${credentials.uniqueIdentifier}_apple_tv`;
-        });
-
-        if(!foundAccessory) {
-          this.debug(`Removing orphaned accessory [${accessory.uid}].`);
-          this.unregisterAccessories([accessory]);
-        }
-    }
-
-    function scanForDevice(deviceConfiguration) {
-        let credentials = appletv.parseCredentials(deviceConfiguration.credentials);
-
-        this.debug(`Scanning for Apple TV with identifier ${credentials.uniqueIdentifier}.`);
-
-        appletv.scan(credentials.uniqueIdentifier).then((devices) => this.onScanComplete(devices, credentials, deviceConfiguration));
-    }
-
-    function onScanComplete(devices, credentials, deviceConfiguration) {
-        if (devices.length) {
-            let device = devices[0];
-
-            this.debug(`Found Apple TV ${device.name} [${device.uid}].`);
-            this.debug(`Attempting to connect to Apple TV ${device.name} [${device.uid}].`);
-
-            device.openConnection(device.credentials).then(connectedDevice => this.onDeviceConnected(deviceConfiguration, connectedDevice), this.onDeviceConnectionFailed);
-        } else {
-            this.debug(`Unable to find Apple TV with identifier ${credentials.uniqueIdentifier}. Please try and pair with your device again.`);
-        }
-
-        this.debug("Scanning complete.");
-    }
-
-    function onDeviceConnected(deviceConfiguration, device) {
-        this.devices.push(new Device(this, deviceConfiguration, device));
-
-        this.debug(`Connected to ${device.name} [${device.uid}].`);
-    }
-
-    function onDeviceConnectionFailed(device) {
-        this.debug(device);
-        this.debug(`Unable to connect to Apple TV.`);
-    }
-
-    this.api.on("didFinishLaunching", this.apiDidFinishLaunching);
 }
 
 Platform.pluginName = "homebridge-appletv-now-playing";
